moved AbstractRecorderConfiguration to org.palladiosimulator.recorderframework.core
