+++ bundles/org.palladiosimulator.recorderframework.core/META-INF/MANIFEST.MF
@@ -5,7 +5,9 @@ Bundle-SymbolicName: org.palladiosimulator.recorderframework.core
 Bundle-Version: 5.2.1.qualifier
 Export-Package: org.palladiosimulator.recorderframework.core,
  org.palladiosimulator.recorderframework.core.config
-Require-Bundle: org.palladiosimulator.measurementframework;bundle-version="5.2.1"
+Require-Bundle: org.palladiosimulator.commons;bundle-version="5.2.1",
+ org.palladiosimulator.measurementframework;bundle-version="5.2.1",
+ org.palladiosimulator.pcm.edp2.measuringpoint;bundle-version="5.2.1"
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Bundle-Vendor: palladiosimulator.org
